Rename *in and *out to simp-in and simp-out
Rename the variables to make it cleare these aren't specvars anymore.
And add a prefix of "simp-" to make it clearer they're for this file
and to make it less likely to collide with any possible specvars in
other files since "in" and "out" are such common words.

Full testsuite passes.