Pass in var to bprog instead of using the special var
Previously `bprog` depended on `var` being special by having it set in
all the callers of `bprog`.

Instead, pass in the var to `bprog` so that we don't need access to
the special `var`.  Name the parameter `bprog-var` to make it unique.
Update all callers of `bprog` to pass in the var.

Full testsuite passes.