Don't declare-top special context
`context` is a specvar defined in db.lisp and is used in exactly one
place in simp.lisp (`$integrate`).  Since simp.lisp is compiled before
db.lisp, we can just add a special declaration in `$integrate` and
remove it from the `declare-top`.

(We should probably move `context` from db.lisp to globals.lisp.  Then
we would need the special declare.  But it's a useful to document this
because `context` doesn't look like a specvar without the "*".)

Full testsuite passes.