Remove specvar ratform
`ratform` was a specvar in simp.lisp that was only used in `disrep`.
Instead of using the specvar, pass in the value via a new arg to
`disrep`, named `disrep-ratform` to make it unique.

This requires finding all callers of `disrep` and adding the second
arg.  But this then means `ratform` doesn't need to be special in
simp.lisp, and now simp.lisp only one specvar with a `declare-top` now.
But simp.lisp still uses `*in-risch-p*` and `context` only in
`$integrate`.  These are the only other specvars in use in simp.lisp.

Full testsuite passes.