Move simple functions into ilt3
The macro `coef` and the functions `lapsum`, `lapprod`, and `expo` are
only used in `ilt3`.  Let's move them into `ilt3` to reduce pollution
of the name space.  For simplicity convert `coef` from a macro to a
function.

Also add a note that `lapsum`, `lapprod`, and `expo` could be replaced
by macros from strmac.lisp.  That would probably be useful.  For
another day.

Full testsuite passes.