Remove specvar q and z by passing them to subroutines
The special vars q and z were used to pass information from `ilt1` to
`ilt2` and `ilt3`.  Instead of using special vars, just pass `q` and
`z` to the functions as needed.  This means `ilt2` needs two new args,
`z` and `q`.  `ilt3` just needs the new arg `q`.  Update callers as
needed.

Note: we also fix a typo where a call to `ilt0` from `ilt3` was not
updated to include the `ils` and `ilt` args.  This is in the `cubic`
branch in `ilt3`.  We should find an example that exercise that branch
of the code.  We should verify the other quadratic branch as well.

Full testsuite passes.