Increase right-binding power of MMINUS from 100 (same as MPLUS) to 134, which is greater than left-binding power of MTIMES (namely 120).
This discourages unnecessary parentheses around the leading term of a "+"
expression when that term has a minus sign.
Possible fix for SF bug #4090: "output formatting: -x-1 has redundant parentheses: (-x)-1"

MMINUS appears in expressions prepared for output by NFORMAT.
The binding power for $- which is recognized by the parser, is unchanged.
Therefore this commit only changes the display of minus expressions,
not input.

There is still some residual strangeness in formatting expressions with minus signs,
which stems from NFORMAT reorganizing (- x)*y to -(x*y).
Printing parentheses for the reorganized expression is actually correct,
because the parser yields (- x)*y, not -(x*y), for the input -x*y.
I believe at this point the focus should shift to NFORMAT.