Rename logptdx to sinint-logptdx and return it from dprog
An examination of the code shows that `dprog` sets `logptdx` so we're
assuming its the main routine for initializing `logptdx`.  So make
`dprog` return two values: the original return value and a second
value of `logptdx`.

Also `fprog` appears to initialize `logptdx` and calls `dprog`, so
modify `fprog` to handle the two return values from `dprog`
appropriately.

Then rename `logptdx` to `sinint-logptdx` to make it distinct from the
specvar `logptdx`.

Finally, in risch.lisp, `rischfprog` calls `dprog` so update the code
to handle the second return value of `dprog` and set it to
`logptdx` (which we didn't rename in risch.lisp).  That's only place
outside if sinint.lisp to call `dprog`.

Full testsuite passes.