Modify cprog not to set parnumer and update code use return value
From what I can tell, `cprog` is the main function that sets the value
of the specvar `parnumer`.  And in all cases it sets it and then
immediately returns.  So, instead, modify all callers of `cprog` to
use the return value to set the local value of `parnumer`.  For
sinint.lisp, we don't want to use `parnumer`, and use the lexical var
`sinint-parnumer` instead.  For risch.lisp, just set `parnumer` to the
return value of `cprog` because we haven't updated risch.lisp to deal
with that yet.

Full testsuite passes.