Rename rootfactor to sinint-rootfactor
It seems `aprog` is the main function to compute and return
`rootfactor`.  Modify `aprog` by renaming `rootfactor` to
`sinint-rootfactor` to make it a lexical var.  `aprog` already
returned `rootfactor`, so thing needed there.

However, all callers of `aprog` need to be updated so that the return
value(s) from `aprog` are actually used to set `rootfactor` (and
`pardenom`) as needed.  For sinint.lisp, this means `dprog` needs to
be updated with a new lexical `sinint-rootfactor` to hold the result
of `aprog`.  It doesn't need to use the specvar `rootfactor` anymore.

In risch.lisp, save both values from `aprog` into the specvars
`rootfactor` and `pardenom`.  (We were already doing `pardenom`.)

Full testsuite passes.