Remove risch-switch1 from tryrisch and rischfprog
The variable `risch-switch1` is initialized to NIL in both `tryrisch`
and `rischfprog` and is not modified.  Let's just remove them and use
NIL in the single use in each function.  We can also get rid of the
now empty `let` in `rischfprog`.

Finally, since `switch1` is no longer used, remove the `declare-top`
declaration for it.

Full testsuite passes.