Add *var parameter to rjfsimp
Add `*var` parameter to `rjfsimp` in nrat4.lisp.  To do this, we also
added a `*var` parameter to callers of `rjfsimp`: `radcan1`, and
`spc0`.  This necessitates updating callers of these functions to add
the `*var` parameter.

In addition, in risch.lisp, we can get rid of the binding for `*var`
that was needed when calling `radcan1`, so remove those now.

Full testsuite passes.