Rename the specvar gamma to risch-gamma everywhere
Rename `gamma` to `risch-gamma` so as not to use a special var when
not needed.

To make this work, had to rework the call to `eval`.  To fix this,
change the places that was setting `denom` to a list to be evaluated.
These were changed into `lambda` functions of no args, and `eval`
becomes `funcall`.

Full testsuite passes.