Make rischform return trigint and hypertrigint (lexical)
Modify `rischform` to use the lexicals `risch-trigint` and
`risch-hypertrigint` instead of specvars `trigint` and
`hypertrigint`.  Then return these lexicals as multiple values.

The only other call to `rischform` updated to handle the multiple
return values.

To make this work, we need to put the original code inside a `labels`
function.  I think this is needed because `rischform` called itself
recursively, but only the last setting of `trigint` and `hypertrigint`
should be used.