Forgot to rename operator to risch-operator in one place
Tests didn't catch this case.  Really need to add a test for it.  We
need to find a test that exercises `intsetup` so that it returns T for
`risch-operator`.  (`intsetup` is full of gotos that it's hard to
figure out path that will lead to that return.  At first glance,
anyway.)