Add nogood parameter to getnfncoeff and dilog
These two functions need the value of `nogood` to be able to return
the correct value because in most situations these functions don't
actually modify the value of `nogood`.

And since we take use the parameter, we can name the parameter
`risch-nogood` to that we use a lexical var instead of a special var.