Rename y to risch-y in intsetup
Make `intsetup` use the lexical var `risch-y` instead of the special
var `y`.  We missed this previously.  We removed the specialness of
`y` so it doesn't matter now, but let's rename it anyway for
consistency with all the other code that previously used `y` and was
renamed.

Full testsuite passes.