Rename var to risch-var in erfarg2
Make `erfarg2` use the lexical var `risch-var` instead of the special
var `var`.

Full testsuite passes.

Since `var` is initialized from `mainvar`, perhaps we can just use
`mainvar`?  But `mainvar` is also special, so let's leave it like this
for now.  But if we replace `mainvar` with a lexical, we can consider
changing it then.