Restructure erfarg2 to pass in mainvar
`erfarg2` initialized `risch-var` from `mainvar` in the `&aux` part of
the lambda-list.  Instead, add `mainvar` as a parameter so we can
replace it with a lexical when the time comes.  While we're at it,
let's also move all the rest of the `&aux` parameters into the `let`.

Update callers appropriately.

Full testsuite passes.