Add expint to functions and return it as needed.
First, in `rischlogpoly`, add `expint` parameter since this function
accesses the special var.  Since it also modifies the value, return
`expint` as the second value of the function.

Update callers appropriately to handle the second return value.

Second, in `rischlogdrog`, add `expint` parameter too and return it as
the third value of the function.

Update callers to handle the new third return value.

Full testsuite passes.