Move body of rischlogeprog into labels function
Use a labels function to implement the guts of `rischlogeprog` where
the labels function takes just a couple of parameters that changes
between the recursive call to itself.  The remaining are closed over.

Add a second return value for `expint`, which isn't used yet by any
caller.

Full testsuite passes.