Add risch-degree arg to rischlogdprog, rischlogpoly
Both `rischlogdprog` and `rischlogpoly` access the special variable
`risch-degree`, so add it as a parameter to the function.

Update callers appropriately.

This is an oversight when we were removing the special var `degree`.
There as a `declare-top` special for `degree` in the middle of the
file that I didn't see, so it got renamed to `risch-degree` by
accident.  Add the parameter to the functions so that we can remove
the special declaration.

Full testsuite passes