Add lians parameter to getfncoeff
`getfncoeff` accesses the special variable `lians` so add it as a
parameter.  Since `getfncoeff` also modifies it, return it as the
another value.

Update callers to handle the extra value.  This also means
`rischexplog` has to initialize it to nil.  `rischlogpoly` calls
`getfncoeff`, but it already initializes `lians` in a `do` loop.

Full testsuite passes.