Remove ratform from declare-top in ratout.lisp
`ratform` is only used in `$partfrac` and does not need to be special
anymore since we've replaced the `ratform` specvar everywhere else
with a lexical var.

Full testsuite passes.

With this change, the last use of `ratform` as a specvar has been
removed as `(find-symbol "RATFORM" "MAXIMA")` shows by returning
`NIL`.

Woohoo!