Update dependencies for risch.lisp
When compiling just risch.lisp, there are warnings about undefined
functions `dosum` and `mtoa` and others.  `dosum` is defined in
asum.lisp and `mtoa` is defined in matrix.lisp.

Update the dependencies so that asum.lisp and matrix.lisp is compiled
before risch.lisp.

Perhaps we should consider breaking up the large "miscellaneous"
module into smaller pieces so we can have a bit finer control over
dependencies.  It seems unfortunate to have to compile all of
"miscellaneous" to satisfy a dependency from risch.lisp.  This also
reduces the chance we'll end up with a circular dependency at some
point.