Rename bigf[12] to ratout-bigf[12]
Rename these special vars so that we use a lexical instead.  These
variables appear to be args to the functions `pgcdp` and `pgcdm` that
want to access them, so there's no reason they can't be lexical.
These two functions are only called from `newgcd`, which immediately
returns after calling them.  And `newgcd` passes in the values for
`bigf1` and `bigf2` so it doesn't matter if the two functions modify
`bigf1` and `bigf2`.

Full testsuite passes.