In package draw, honor fill_density in errors, polygon, ellipse, bars, explicit, and region.
In errors and bars, the default is to draw the object empty (no fill).
For the other objects, the default is density = 1 (solid fill).

In INI-GR-OPTIONS, set default value for fill_density to NIL,
and then determine how to handle it for each kind of object.
For errors and bars, emit 'fc 0' (fillcolor 0) in Gnuplot output.
For other objects, omit output altogether when fill_density is null.