Manually revert cde0c181
Commit [cde0c181] updated dependencies for risch.lisp to get `dosum`
and `mtoa` (at least) to be compiled before risch.lisp.  However, this
causes test failures with ecl.  I just did a git bisect to find that
[cde0c181] is the first bad commit.

I didn't investigate why this causes test failures.  I'm guessing some
of the files declare a variable special, which also happens to be the
name of some variable in risch.lisp, so things are messed up.  (If so,
this is yet another reason to remove as many special variables as
possible.)

Reverting the commit manually allows ecl to pass the testsuite without
errors, as expected.