Rename `v` to `ratout-v` in `wtptimes1`
Do the renaming so that we don't use the special variable `v`.  To
make this work also let `wtptimes3` close over the new local variable
`ratout-v` since `wtptimes3` modifies the value and is also called in
a loop, so updated values of `ratout-v` get carried to later calls.

Remove the special declarations in `wtptimes3` and `wtptimes1` too.

Full testsuite passes.