Move pgath3 into pmodcontent in pgcdp
`pgath3` is only called from `pmodcontent` in `pgcdp` so move it as a
`labels` funtion in `pmodcontent`.  This is useful becase pgath3
references the special variable `*res` that is set in `pmodcontent`.

This could probably be done in a different way so as not to make
`pmodcontent` a huge function, but this is simpler because we can have
`pgath3` close over the `*res` in `pmodcontent`.  It's more
complicated if we don't do that because `pgath3` modifies `*res`.

Full testsuite passes.