Fix potential error in wtptimes1
Previously we had something like:
```
(defun wtptimes1 (ratout-x ratout-y xweight)
  (let (ratout-v ratout-u*)
    (labels ((wtptimes2 (ratout-y)
              ...)
	     (wtptimes3 (ratout-y)
	      ;; Accesses ratout-v and ratout-u*
	      ))
      (prog (ratout-v)
        ;; call wtptimes3
	))))
```
So `wtptimes3` modified `ratout-v` from the `let`, not the `prog`.
This seems wrong, but the testsuite passed.  Perhaps there's no test
that exercises that path?

Anyway, get rid of `ratout-v` from the `prog` so that the the captured
`ratout-v` from the `let` is used as we had originally intended.

Full testsuite still passes.