Configure sets the texi initfile for xmaxima
To create the correct html file names, we need to use the correct texi
init file, which is set by configure based on the version of makeinfo
being used.  (See build_html.sh.in for how this is done.)

Thus, update interface/xmaxima/doc/Makefile.am to have configure fill
in the desired file name for the init file.