Declare context special in laplac; remove unspecial in risch
Previously there were compiler warnings about `context` being
undefined in laplac.lisp.  This happens because laplac.lisp is
compiled after risch.lisp and risch.lisp had an unspecial declaration
for `context` at the end of the file.

Since `context` is a `defvar` in db.lisp, let's fix this by removing
the unspecial declaration from risch.lisp and adding a special
declaration in laplac.

This fixes the warning in laplac.lisp.

Perhaps a better fix is to move the `defvar` for `context` in db.lisp
to globals.lisp?