Restore $pfet
We had previously commented it out because it wasn't documented.
Let's restore it for now for backward compatibility.

In doing so, move the functions `sssqfr` and `pfet1` into `$pfet`
since they're only called from `$pfet`.  We can also get rid of the
`declare-top` stuff around `$pfet` as they're not needed: `varlist`,
`$factorflag` are globals from globals.lisp; `m`, and `v` don't need
to be special because no function is referencing either without it
being a parameter.