Move floatproperties.lisp to globals.lisp
We want to have the float properties as part of the core of Maxima, so
move floatproperties.lisp to globals.lisp.  These constants and
functions are fundamental properties of Maxima flonums so they should
be part of the core.

As part of the move, use the `:properties` feature of `defmvar` to
disallow setting of these values instead of directly setting the
`assign` property.