Support clisp which doesn't have denormals.
Adjust the value of `+most-negative-normalized-float-exponent+` for
clisp (and other lisps where `$least_positive_float` and
`$least_positive_normalized_float` are the same).

Update tests to use `$least_positive_float` as the result instead of
using an explicit numerical value that is wrong when Lisp doesn't
support denormals.

With these changes, clisp passes all the tests except the one
involving `unit_in_last_place(float((0+2^52)/2^1073))` because that
assumes denormals are supported.