Add support for deprecated variables.
This is based on a proposal from Stavros.

Update `defmvar` to recognize a `:deprecated-p` option that takes a
string that is the deprecation message.  This overrides any other
options.

When a variable is deprecated, it is not resettable.  It is unbound.
The 'bindtest property is set.  The 'assign property is set to
'neverset so it can't be changed.  The option string is added to
`*bindtest-messages*`.

We also add the global variable *bindtest-messages* which is an alist
whose key is the deprecated variable and the datum is the message.

mlisp.lisp is updated so that an string from `*bindtest-messages*` can
be used instead of the default when accessing an unbound variable.
This works because we set the 'bindtest property to t.

Finally, make `$largest_float` and `$largest_negative_float`
deprecated using the new `defmvar` option `:deprecated-p`.