Simplify the supplied deprecation message.
Make `:deprecated-p` accept a string.  Then we'll prepend "Deprecated
variable `~M':" because the bindtest wants on "~M" for the variable
name.  This way, we're always guaranteed to have the "~M".  The
user-supplied string should not have any `~` commands.

Update `$largest_float` and `$largest_negative_float` to use the new
format for the deprecation message.