Rework share/orthopoly/test_orthopoly.mac into (test case, result) pairs, as with other test suite files. Rename test cases to rtest_orthopoly.mac and move plotting commands into a separate script, plot_orthopoly.mac. test_orthopoly.mac is removed in this commit.
Minimal changes were made in order to convert test_orthopoly.mac.
In particular, the function check_zero_list returns the name of the
test if the test passes; that is unchanged, I did not consider
whether there is some other return value which could be more effective.

All the orthopoly tests are the same, none were added or taken away.