Change how deprecated vars work
Previously, deprecated variables produced an error when used.  This is
a weird interpretation of "deprecation", which usually means something
is deprecated but still available.

So, instead, when the variable is first used, we print a warning but
continue with the expected value.  This requires updating `defmvar` to
handle this.  Update float-properties.lisp for the new `defmvar`.

Update mlisp.lisp to handle this too.  We check if the 'bindtest
property has the value `:deprecated`.  If so, we look up the symbol in
`*bindtest-messages*`.  If found, we get the deprecation message to
use along with the value.  A warning is printed with the message, and
the variable is set to the value.  Hence, we never reach the bindtest
part of the code again because the variable is now bound.