Allow front-ends to provide additional info for bug_report()
Until now bug_report() did tell the user to report bugs to the maxima
project. But it did not tell the user that if the GUI or the way maths
is displayed is broken that should be reported to the front-end
manufacturer instead, if a front-end is used. It also didn't tell
how to reach that front-end manufacturer.

Squashed commit of the following:

commit a9a121bf47e267ed285df160c9077064ea88981b
Author: Gunter Königsmann <gunter@peterpall.de>
Date:   Sun Mar 26 16:13:43 2023 +0200

    The integrated front-ends now report that they are part of maxima

commit 99b38354c39d7060af81813092a8fb6fec7afbca
Author: Gunter Königsmann <gunter@peterpall.de>
Date:   Sun Mar 26 16:06:32 2023 +0200

    Added a ChangeLog entry with the new variable

commit 051efb4af9ac8a63195cad6565480a8e93b85650
Author: Gunter Königsmann <gunter@peterpall.de>
Date:   Sun Mar 26 16:04:08 2023 +0200

    Provide the maxima frontend with a way to provide bug report info