Fix incorrect --no-init that wasn't working
--no-noinit is supposed to disable loading of the init files.
However, it wasn't working because the user init files were loaded
before the command-line args were processed.  Therefore, load the user
init files after processing the command-line args.

Also, print out the init files that are loaded.  This is useful to see
that we've loaded the user init files (and where).  Also useful if we
expected an init file to be loaded and wasn't or the wrong init file
was loaded.

This includes any files loaded via the -p option.  This is useful to
know exactly which file was loaded since the -p option uses
file_search to find the file to load.