Greedier logic for autodetecting which lisps to compile with
If the user hasn't specified a lisp to compile maxima with and
we already have autodetected that more than one suitable lisp
is installed on the system there is no good reason to just pick
one lisp by chance and to compile using that one.
Instead we should compiling using all available lisps in that
case.

If the user has explicitely chosen which lisps to compile on
that choice is respected and no additional lisps are used -
just like ./configure did until now.

One side-effekt this change has is that it makes things easier
for packagers as it means that even if maxima is compiled with
more than one lisp no special ./configure args are needed.