Remove a no-more-needed warning
We still warned about the problem that clisp, if stdout is bound to
a network socket on (force-output) or (clear-output) sometimes sends
network packets in the wrong order.

Since we now only use (finish-output) for that purpose the
problem currently doesn't occur and we don't need to warn
about it.

One possible TODO might be to automatically check for
(force-output) and (clear-output) at compile time in order
to warn if someone uses the broken commands. But do we care
so much about bugs in foreign software that we need to?