Compute subdirs of .maxima and share at runtime
On startup we find all subdirectories if *maxima-userdir*
and *maxima-sharedir* so that if anything was added, we automatically
have access to them.

For *maxima-sharedir*, this is useful because we don't have to
maintain share-subdir.lisp.

For *maxima-userdir*, this is useful if we add new packages into
subdirectories.  Then we'll automatically have access to them when
maxima starts.