Add third arg (For TeX) to m4_math for expressions with "\lt".
Generating the pdf documentation failed (with texinfo-7.0.3)
when the first arg of macro m4_math contains "\"lt".  Add third
arg - for TeX - that uses "<".

I don't understand if this is a if this is a texinfo version problem or
some other toolchain issue.  rtoy suggested this solution and
it works for me.  Committing to get the build working.