Need to add @node for two @sections
The sections "General global variables" and "Variables related to the
modular test" were marked as sections, but did not have a
corresponding "@node".  Hence the text index did not include these
subsections like all the other subsections in the manual.

With this change, the html index is now exactly the same as the text
index.