Move html verification code to macdes.lisp
init-cl.lisp is not always compiled so move the verification code to
macdes.lisp which is compiled.  Hopefully this will make it a little
master.

Temporarily added some calls to time to see where the cost is.  For
cmucl, 0.4 sec is spent in getting the html index and 0.4 sec is spent
in computing the set difference.  Everything else is essentially 0.
Presumably most of the time in getting the html index is in pregexp.
I guess set differences of long lists is slow.