Use equal for the set-difference test instead of string=
This is a big speed up for clisp which has special methods for
set-difference when the test is one of the tests allowed for
hash-tables.  On my test, the time to do both calls to set-difference
changed from 0.28 sec to 0.14 sec.

This would be useful for any other lisps that do similiar
optimizations.