Fix issues replacing "-1" with " <1>"
We were failing to convert things like "Logical-conjuction-1" to
"Logical conjunction <1>".

First, we need to convert dashes not followed by a digit with spaces
to get "Logical conjuction-1".  Then convert "-n" with " <n>" so we
get "Logical conjunction <1>", which is the correct info item.

This also means we don't need to do this conversion in macdes.lisp to
find the desired html entry.  We can use it as is.