Operate on hash tables directly for set-difference
Instead of converting the hash tables for the html entries and the
text entries to lists for set-difference, operate on the tables
directly.

Well, almost.  We create a new hashtable of text entries
because *text-topics* is a hashtable of hashtables of the entries.

This is a huge speed up on clisp and cmucl.  Time is now on the order
of 1 millisecond instead 1/4 sec or more.