Don't specify more elements for subseq than there are elements.
We were doing something like `(subseq diffs 0 20)` even when `diffs`
did not have 20 elements.  for some reason this causes clisp to signal
an error and then produce a segfault.  That shouldn't happen in clisp,
but we also shouldn't specify too many elements.

Thus, take the min of 20 and the list length.