Add $verify_html_index
Add the Maxima function $verify_html_index so it can be run from the
repl for debugging; it is otherwise undocumented on purpose.  The
optional timep arg will print out information (from `time`) to show
how much time is spent during verification.  This is meant for
debugging too to see how much time is spent, if startup seems to have
gotten slower.