Clean up for building html index
Update comments in Makefile.

Don't try to verify html index when building the index because it
won't exist on a clean checkout.

But after it's done, run maxima again to verify the index.  We don't
currently do anything with the results and depend on the builder to
notice if something went bad.  Perhaps that's ok.  Unless the html
index is really broken, having some html help is better than a broken
build.