Fix up path to "maxima_toc.html"
Need to merge the file name "maxima_toc.html" with the directory so we
can actually find the file.  This only matters if you run
build_html_and_dump_html_index with a path that is not the current
directory.

Return true if build_html_and_dump_html_index has more than 1 entry in
the table.  Otherwise return false.