Fix #4149: Move fnindex and vrindex before the definition
In the HTML docs, the links for the @fnindex and @vrindex items would
be placed after the definition line.  This is pretty annoying because
you would have to scroll up to find the definition line.  By moving
the index commands before the definition, the links now show the
definition line.  No change to what the text help shows.

I did not test every link, but did test quite a few and they are in
the right place for both html and text.