Windows installer: Update the build instructions.
I did set up a Github repository, where I use the Github CI infrastructure
to show how the installer is build.
And now I can exactly tell, which packages must be installed (on their
standardized build machines), so that should be reproducible by everyone.

There seem some small bugs in the Ubuntu dependencies, but now it works.

Found a workaround how to build a Maxima installer in a non-interactive
shell, cronjob or the CI build machines. And documented that.