Don't rename html files to maxima_nnn.html
Testing with texinfo 6.8.

Remove the stuff in texi2html.init.in that renamed the output files to
maxima_nnn.html.  We now let makeinfo choose the names for us.

Now the ToC is in index.html and the function index is in
Function-and-Variable-Index.html.  We don't need to search for this
file anymore.

Update build-html-index.lisp with new regexes to match the items we
want.

Had to update display-html-help to handle the new links.

Validation shows that the html index exactly matches the text index.
A random sampling of topics takes me to correct places in the browser.