Fix up regex for function and variable index entries
We were getting the wrong id links and using the wrong id.  The id
links (item-id) is everything after "#" so it includes "index-".  The
id is everything after "#index-".

And we want to return both this id and item-id so that add-entry puts
the right things into the final index file.