Handle right quotation marks for texinfo 7.0.2 as well
Texinfo 7.0.2 behaves like 7.0.3 with using right single quotation
marks in the info file, but plain apostrophes for the html file.
Entries need to be updated for this.

Also took the opportunity to use a function to convert the major,
minor, and patch values to an integer that can be used for comparing
versions.

Finally, fixed a typo in a comment.