Bug #4153:  rectform(atan2(1,1+%i)) signals error
`risplit` calls `logarc` with `$atan2`, but `logarc` is looking for
`%atan2`.  So, just call `logarc` with `%atan2` (instead of `(caar
l)`, because we already tested that `(caar l)` is `$atan2`.

But the second arg to `logarc` was given as the first arg of `atan2`.
We really want a list of all the args, so use `(rest l)` instead of
`(cadr l)` for the second arg.

The test case now returns atan(2)/2, as expected.