Use def-simplifier for bern
Use def-simplifier to define the simplifier for bern instead of using
the hand-written simpbern.  This guarantees that we have all the
properties set up correctly.

This also means replacing all calls to `$bern` with `(ftake '%bern
...)`.

For powerseries, we need to replace the expansion with `%bern` instead
of `$bern`.

With these changes the testsuite still passes.