Update give-up in def-simplifier to take optional args
Modify give-up in def-simplifier to take optional args, defaulting to
the args of the simplifier.  This allows simplifiers like the
simplifier for genfact to specify new args before giving up.  The new
args take the floor of the second arg if possible, so genfact(a,3/2,c)
will return genfact(a,1,c).  Not sure why that is, but that's how it
currently works.

Testsuite passes.