Handle languages better
build-html-index needs handle more html entities.  We added some for
German.

To handle the languages, add a hashtable to map the language to the
possible file name containing the function and variable index.  If
that doesn't work, look for a string that is the title of the function
and variable index section.

build_and_dump_html_index gets a second arg for the language.  This
should be a string like "de" (German) or "" (English).

Update common-lang.mk to have the right dependencies and sets
the envvar MAXIMA_LANG_SUBDIR so maxima can find the appropriate
maxima-index-html.lisp file.  This is needed for verification of the
html index.