Replace more html entities with corresponding character
When parsing the HTML file, we need to replace html entities with the
corresponding character to match text info file.  This was done by
looking at the verification output to see what characters might be
converted to an HTML entity.

Update the hashtable for the index file name and title to include pt
and es.

Add some interface functions for extracting the index file name and
title.

If lang is not given to build_and_dump_html_index, default to "",
which means English.