Sort the result of DIRECTORY
We need to sort the result from DIRECTORY because that's how the old
search list (for the share directories) were sorted.  If we don't, we
might get the wrong file if there are files with the same name.  For
example, there are two files named "format.lisp" in two different
directories.

Add *debug-new-file-search* to enable some debugging prints for
new-file-search.