Clean up search list extensions
Remove some extensions:
  * No longer use "lsp" for lisp files
  * Only recognize "dem" for demo files
  * Only recognize "usg" for usage files

Modify build-search-list to take an alist whose car is the directory
and the cdr is the list of extensions to search for.  This allows each
directory to allow specific extensions.  In particular, files in the
share dir only need "lisp" and "mac".  There are now "lsp" or "wxm"
files there.