Fix #4158: Make float(%i) return 1.0*%i
float(li[2](%i)) was  not resulting in a numerical value because
float(%i) is %i.  The simplifier sees the arg as %i and doesn't
numerically evaluate.

To fix this, modify $float to convert %i to be 1.0*%i.  Need to do the
same for bfloat.

This also fixes the issue where li[2](%i),numer was causing a lisp
error.  This works because we call $float before calling
bigfloat::li2numer to compute the numerical result.