Support makeinfo 5.1 when building html index
With makeinfo 5.1, the ToC is in maxima_0.html, not index.html, so
search we need to see if either of them exist.  Of course, we try
index.html since that's what 6.8 and later create.

Also, makeinfo doesn't create a separate file for the Function and
Variable Index, we have to search for it.  The pattern for the title
that we're looking for is slightly different from what we had, update
the regexp for it.

With this change, the index can be created when using makeinfo 5.1.
However, only 2817 entries are found out of the expected 3077.  The
missing entries include

 "Airy Functions" "Arithmetic operators" "Arrays" "Assignment
 operators" "Bessel Functions" "Characters" "Combinatorial Functions"
 "Command line options" "Comments" "Constants" "Definitions for IFS
 fractals" "Definitions for Koch snowflakes" "Definitions for Peano
 maps" "Definitions for complex fractals" "Directory operations"
 "Documentation" "Environment operations" "Error Function" "Error
 Messages" "Examples for cobyla"... plus 240 more