Fix #4156: Float of generalized_lambert_w
When evaluating `float(generalized_lambert_w(-1,x))`, the simplifier
sees the args -1.0, and float(x).  The code currently only handles the
case where k is an actual integer, not a float that is an integer.

To fix this, test to see if k is numerically equal to an integer (via
truncate).  If it is, set k to be the integer.