Fix #4156: float of generalized_lambert_w
In `float(generalized_lambert_w(k,x))`, the simplifier for
`generalized_lambert_w` sees the args `float(k)` and `float(x)`.  The
current simplifier only simplifies the result if the first arg is an
integer, not a float that is equal to an integer.

To fix this, treat a float (or bfloat) value as if it were an
integer.