Handle k arg better.
The previous version was incorrectly handling the k arg.

First, to make things a little easier, use def-simplifier to define
the simplifier for generalized_lambert_w.  Then we can get rid of the
handwritten code for $generalized_lambert_w and also remove manual
property settings for noun/verb/operator stuff.

Second, use (complex-float-numerical-eval-p k x) to test if we can
numerically evaluate the function.  If true, then we know k is an
numerical value of some sort.  Then we can check to see if k is equal
to an integer and use the integer value to call the appropriate
routines.  If not, then we bail.