Change usae of m4_math and friends
Because m4_math expands into things like `@html`/`@end html`, m4_math
needs to make sure `@html` is at the beginning of the line.
Previously, m4_math would insert a new line at the beginning.
However, this sometimes caused unexpected new lines to appear in the
output.

We changing the macro so that the new line is not inserted anymore.
That means m4_math and friends MUST be on a line by itself.

All the uses of m4_math and friends have been updated so that m4_math
is on a line by itself.  I checked the pdf and html resuls, and they
look good, without extraneous newlines or extra space.