Define m4_mathpunc to allow trailing punctation
m4_mathpunc is like m4_math, except the first arg is punctuation that
is appended to each arg.  This works better then having the
punctuation at the end of m4_math because there's generally too much
space between the formula and the punctuation when done that way.  By
adding the punctuation at the end of the formulas, there's less space
so it looks better, especially for periods and commas.