Use absolute path to texi2html.init for make
common-lang.mk is included in each lang subdirectory and the Makefile
has a dependency on texi2html.init.  Use an absolute path to this
file (in doc/info/texi2html.init).  This shows up in building the ja
docs, which complains that texi2html.init doesn't exist which is true
because the ja directory doesn't have a copy.